+++ src/Controllers/LancamentoManualCaixaController.php
@@ -1,6 +1,7 @@
 <?php
 namespace App\Controllers;
 use App\Models\MovimentoCaixaModel;
+use App\Models\LancamentoManualCaixa; 
 class LancamentoManualCaixaController
 {
     public function index()
@@ -12,7 +13,12 @@ class LancamentoManualCaixaController
     {
         if ($_SERVER['REQUEST_METHOD'] === 'POST') {
             $lancemento = new LancamentoManualCaixa();
-            $lancemento->data_lancamento = $_POST['data_lancamento'];
+            
+            // CORREÇÃO PRINCIPAL: 
+            // 1. Usa a propriedade correta do Modelo (data_movimento)
+            // 2. Assume que o nome do campo no formulário HTML também foi corrigido
+            $lancemento->data_movimento = $_POST['data_movimento'];
+            
             $lancemento->descricao = $_POST['descricao'];
             $lancemento->valor = $_POST['valor'];
             $lancemento->tipo = $_POST['tipo'];
